+++ packages/plugin-legacy/src/index.ts
@@ -1,8 +1,8 @@
 /* eslint-disable node/no-extraneous-import */
-import path from 'path'
-import { createHash } from 'crypto'
-import { createRequire } from 'module'
-import { fileURLToPath } from 'url'
+import path from 'node:path'
+import { createHash } from 'node:crypto'
+import { createRequire } from 'node:module'
+import { fileURLToPath } from 'node:url'
 import { build, normalizePath } from 'vite'
 import MagicString from 'magic-string'
 import type {
@@ -625,7 +625,8 @@ async function buildPolyfillChunk(
     plugins: [polyfillsPlugin(imports, excludeSystemJS)],
     build: {
       write: false,
-      target: false,
+      // if a value above 'es5' is set, esbuild injects helper functions which uses es2015 features
+      target: 'es5',
       minify,
       assetsDir,
       rollupOptions: {